updated to long time zone name

+++ src/course-home/outline-tab/SequenceLink.jsx
@@ -49,7 +49,7 @@ const SequenceLink = ({
             day="numeric"
             month="short"
             year="numeric"
-            timeZoneName="short"
+            timeZoneName="long"
             value={due}
             {...timezoneFormatArgs}
           />
@@ -71,7 +71,7 @@ const SequenceLink = ({
             day="numeric"
             month="short"
             year="numeric"
-            timeZoneName="short"
+            timeZoneName="long"
             value={due}
             {...timezoneFormatArgs}
           />
